+++ src/clients/Messages.ts
@@ -1,6 +1,4 @@
-import { InstanceOptions } from '../HttpClient'
-import { inflightUrlWithQuery } from '../HttpClient/middlewares/inflight'
-import { forWorkspace, IODataSource } from '../IODataSource'
+import { AppClient, inflightUrlWithQuery, InstanceOptions } from '../HttpClient'
 import { IOMessage } from '../service/graphql/schema/typeDefs/ioMessage'
 import { IOContext } from '../service/typings'
 
@@ -12,18 +10,12 @@ interface Locales {
   [lang: string]: Locale
 }
 
-export class Messages extends IODataSource {
-  protected httpClientFactory = forWorkspace
-  protected service = 'messages.vtex'
-
-  constructor(vtex: IOContext, options: InstanceOptions) {
-    super(vtex, options)
+export class Messages extends AppClient {
+  constructor(vtex: IOContext, options?: InstanceOptions) {
+    super('vtex.messages', vtex, options)
   }
 
   public translate = (to: string, data: IOMessage[]): Promise<string[]> => this.http.get('/_v/translations', {
-    headers: {
-      Authorization: this.context!.authToken,
-    },
     inflightKey: inflightUrlWithQuery,
     metric: 'messages-translate',
     params: {
@@ -33,10 +25,6 @@ export class Messages extends IODataSource {
   })
 
   public saveTranslation = (data: Locales): Promise<void> => this.http.post('/_v/translations/save', data, {
-    headers: {
-      Authorization: this.context!.authToken,
-    },
     metric: 'messages-save-translation',
   })
 }
-