+++ CHANGELOG.md
@@ -7,6 +7,14 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+## [3.7.0] - 2019-04-25
+
+### Changed
+- Add LicenseManager client
+- Add Session client
+- Restructure HttpClient
+- Add HttpClient subclasses
+
 ## [3.6.1] - 2019-04-24
 
 ## [3.6.0] - 2019-04-24